Fix a bug where the apple bin image wouldn't return to the middle of the screen after the animation of the bin swap
Update pickingMode and showBinFullnessLevel functions to change the way bin images are being displayed. Also update nextBin function to make sure that the picture of a full bin will be visible as the bin animation runs, moving the bin off screen, and only get replaced with the picture of an empty bin when the "new" bin is being brought in from the other side of the screen.